Add more fine tune examples

+++ cmd/gitemoji/install.go
@@ -11,9 +11,6 @@ import (
 	"github.com/spf13/cobra"
 )
 
-//go:embed fine-tunings.jsonl
-var fineTuningsRaw string
-
 func newClient() *openai.Client {
 	const keyEnv = "OPENAI_API_KEY"
 	key := os.Getenv(keyEnv)
@@ -23,31 +20,40 @@ func newClient() *openai.Client {
 	return openai.NewClient(key)
 }
 
+const openAIFineTuneName = "gitemoji"
+
 func installCmd() *cobra.Command {
-	return &cobra.Command{
+	var printFineTune bool
+	cmd := &cobra.Command{
 		Use: "install",
 		Run: func(cmd *cobra.Command, _ []string) {
+			if printFineTune {
+				err := generateFineTunings(os.Stdout)
+				if err != nil {
+					flog.Fatalf("gen fine tuning: %v", err)
+				}
+				os.Exit(0)
+			}
+
 			client := newClient()
 			tmpFi, err := os.CreateTemp("", "gitemoji")
 			if err != nil {
 				flog.Fatalf("create temp file: %v", err)
 			}
 			defer os.Remove(tmpFi.Name())
-
-			err = os.WriteFile(tmpFi.Name(), []byte(fineTuningsRaw), 0o644)
+			err = generateFineTunings(tmpFi)
 			if err != nil {
 				flog.Fatalf("write temp file: %v", err)
 			}
-
-			const fileName = "gitemoji"
 
 			files, err := client.ListFiles(cmd.Context())
 			if err != nil {
 				flog.Fatalf("list files in openai: %v", err)
 			}
 
+			// Delete old file if they exists.
 			for _, file := range files.Files {
-				if file.FileName == fileName {
+				if file.FileName == openAIFineTuneName {
 					err = client.DeleteFile(cmd.Context(), file.ID)
 					if err != nil {
 						flog.Fatalf("delete file in openai: %v", err)
@@ -57,7 +63,7 @@ func installCmd() *cobra.Command {
 			}
 
 			apiFile, err := client.CreateFile(cmd.Context(), openai.FileRequest{
-				FileName: fileName,
+				FileName: openAIFineTuneName,
 				Purpose:  "fine-tune",
 				FilePath: tmpFi.Name(),
 			})
@@ -66,6 +72,7 @@ func installCmd() *cobra.Command {
 			}
 
 			ftm, err := client.CreateFineTune(cmd.Context(), openai.FineTuneRequest{
+				// Model:        "text-davinci-003",
 				Model:        "davinci",
 				TrainingFile: apiFile.ID,
 			})
@@ -73,7 +80,7 @@ func installCmd() *cobra.Command {
 				flog.Fatalf("create fine-tune in openai: %v", err)
 			}
 
-			flog.Infof("fine-tune model %+v created", ftm)
+			flog.Infof("fine-tune model %+v created, waiting to finish training", ftm)
 
 			for r := retry.New(time.Second, time.Second*5); r.Wait(cmd.Context()); {
 				ftm, err = client.GetFineTune(cmd.Context(), ftm.ID)
@@ -83,10 +90,6 @@ func installCmd() *cobra.Command {
 				switch ftm.Status {
 				case "succeeded":
 					flog.Successf("fine-tune model %+v has succeeded", ftm)
-					err = fineTuneModelID.Write([]byte(ftm.ID))
-					if err != nil {
-						flog.Fatalf("write fine-tune model id: %v", err)
-					}
 					return
 				default:
 					flog.Infof("fine-tune model %q is %q", ftm.ID, ftm.Status)
@@ -94,4 +97,7 @@ func installCmd() *cobra.Command {
 			}
 		},
 	}
+
+	cmd.Flags().BoolVar(&printFineTune, "print-fine-tune", false, "print fine-tune file to stdout and exit")
+	return cmd
 }
